Add resource package
